+++ src/index.js
@@ -1,4 +1,4 @@
-const CommandStructures = require('./structures/command')
+const CommandStructures = require('./structures/command');
 
 module.exports = {
   // Command Structures
@@ -17,7 +17,7 @@ module.exports = {
   Reddit: require('./utils/Reddit.js'),
   Constants: require('./utils/Constants.js'),
   MiscUtils: require('./utils/MiscUtils.js'),
-  CanvasTemplates: require('./utils/CanvasTemplates.js'),
+  CanvasTemplates: require('./utils/CanvasTemplates'),
   PermissionUtils: require('./utils/PermissionUtils.js'),
   BlacklistUtils: require('./utils/BlacklistUtils.js'),
