+++ src/locales/en-US/commands.json
@@ -631,5 +631,11 @@
     "commandUsage": "<user to slap>",
     "noMention": "You need to mention someone",
     "success": "{{- _author}} slapped {{- slapped}} BAKA :<"
+  },
+
+  "buttslap": {
+    "commandUsage": "<user to slap>",
+    "noMention": "You need to mention someone",
+    "sucess": "{{- _tom}} hit the butt of {{- _kettin}}"
   }
 }
