+++ src/utils/CanvasTemplates/Triggered.js
@@ -1,0 +1,43 @@
+const Constants = require('./Constants');
+const { createCanvas, Image } = require('canvas');
+const GIFEncoder = require('gifencoder');
+
+class Triggered {
+    static async render (user) {
+        const WIDTH = 256
+        const HEIGHT = 310
+    
+        const IMAGE_ASSETS = Promise.all([
+          Image.from(Constants.TRIGGERED_LABEL_PNG, true),
+          Image.from(user.displayAvatarURL.replace('.gif', '.png'))
+        ])
+    
+        const encoder = new GIFEncoder(WIDTH, HEIGHT)
+        encoder.start()
+        encoder.setRepeat(0) // Repeat
+        encoder.setDelay(50) // 50ms delay between frames
+    
+        const canvas = createCanvas(WIDTH, HEIGHT)
+        const ctx = canvas.getContext('2d')
+    
+        const [ triggeredLabel, avatarImage ] = await IMAGE_ASSETS
+    
+        const AVATAR_RANDOM_MAX = 20
+        const LABEL_RANDOM_MAX = 10
+        const random = (max) => Math.floor(Math.random() * max) - max
+        for (let i = 0; i < 8; i++) {
+          ctx.clearRect(0, 0, WIDTH, HEIGHT)
+          ctx.drawImage(avatarImage, random(AVATAR_RANDOM_MAX), random(AVATAR_RANDOM_MAX), WIDTH + AVATAR_RANDOM_MAX, HEIGHT - 54 + AVATAR_RANDOM_MAX)
+          ctx.fillStyle = '#FF000033'
+          ctx.fillRect(0, 0, WIDTH, HEIGHT)
+          ctx.drawImage(triggeredLabel, random(LABEL_RANDOM_MAX), HEIGHT - 54 + random(LABEL_RANDOM_MAX), 256 + LABEL_RANDOM_MAX, 54 + LABEL_RANDOM_MAX)
+          encoder.addFrame(ctx)
+        }
+    
+        encoder.finish()
+    
+        return encoder.out.getData();
+      }
+}
+
+module.exports = Triggered;
