+++ src/utils/CanvasTemplates/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    ButtSlap: require('./ButtSlap.js').render,
+    
+    tiggered: require("./Triggered.js").render,
+    profile: require("./Profile.js").render,
+    nowPlaying: require("./NowPlaying.js").render,
+    leaderboard: require("./LeaderBoard.js").render
+}
